Refactor aggregation methods
